+++ export.c
@@ -1568,6 +1568,8 @@ size_t print_https_common_name(struct printbuf *kafka_line_buffer,
     https_command_name_nbar_id,sizeof(https_command_name_nbar_id));
 }
 
+#endif /* HAVE_CISCO_URL */
+
 #ifdef HAVE_UDNS
 
 static size_t print_dns_obtained_hostname(struct printbuf *kafka_line_buffer,
@@ -1585,8 +1587,6 @@ static const char *get_direction_based_target_hostname(struct flowCache *flowCac
   return flowCache->address.target_name ? flowCache->address.target_name :
     flowCache->address.target_name_cache ? flowCache->address.target_name_cache->name : NULL;
 }
-
-#ifdef HAVE_UDNS
 
 static const uint8_t *get_src_ip(struct flowCache *flowCache) {
   return (uint8_t *)flowCache->address.src;
@@ -1634,7 +1634,6 @@ const uint8_t *get_direction_based_target_ip(struct flowCache *flowCache) {
     dst_ip : src_ip;
 }
 
-#endif /* HAVE_UDNS */
 
 /// @TODO difference client/target src/dst
 size_t print_client_name(struct printbuf *kafka_line_buffer,
@@ -1653,6 +1652,427 @@ size_t print_target_name(struct printbuf *kafka_line_buffer,
     get_direction_based_target_hostname(flowCache));
 }
 
-#endif
+#endif /* HAVE_UDNS */
 
+size_t printNetflowRecordWithTemplate(struct printbuf *kafka_line_buffer,
+    const V9V10TemplateElementId *templateElement,
+    const void *vbuffer, const size_t real_field_len,
+    const size_t real_field_len_offset, struct flowCache *flowCache) {
+  const uint8_t *buffer = vbuffer;
+  #ifdef CISCO_URL
+  #if 1
+  //static __thread char url[2048] = {'\0'};
+  //static __thread int social_user_printed = 0; // useful for coordinate host+url and referer social user printing
+  #endif
+  #endif
+  const int start_bpos = kafka_line_buffer->bpos;
+  int value_ret=0;
+  if(0!=strcmp(kafka_line_buffer->buf,"{")){
+    printbuf_memappend_fast(kafka_line_buffer,",",strlen(","));
+  }
+  printbuf_memappend_fast(kafka_line_buffer,"\"",strlen("\""));
+  printbuf_memappend_fast(kafka_line_buffer,templateElement->jsonElementName,
+    strlen(templateElement->jsonElementName));
+  printbuf_memappend_fast(kafka_line_buffer,"\":",strlen("\":"));
+
+  if (templateElement->quote) {
+    printbuf_memappend_fast(kafka_line_buffer,"\"",strlen("\""));
+  }
+
+  if(likely(NULL!=templateElement->export_fn)){
+    value_ret = templateElement->export_fn(kafka_line_buffer,buffer,real_field_len, real_field_len_offset, flowCache); // at the moment
+  }
+
+
+  {
+    // @TODO Move all this .
+    const size_t bufsize = 1024;
+    char buf[bufsize];
+    buf[0] = '\0';
+
+    switch(templateElement->templateElementId)
+    {
+      case CISCO_HTTP_SOCIAL_MEDIA_YT:
+      case CISCO_HTTP_SOCIAL_USER_FB:
+      case CISCO_HTTP_SOCIAL_MEDIA_TT:
+      case CISCO_HTTP_SOCIAL_USER_TT:
+      case CISCO_HTTP_SOCIAL_MEDIA_IG:
+      case CISCO_HTTP_USER_AGENT_OS:
+        if(real_field_len < 6)
+          break;
+        snprintf(buf,min(sizeof(buf),(unsigned) real_field_len-6+1),"%s",buffer+real_field_len_offset+6);
+        break;
+      default:
+        // Do nothing
+        break;
+    };
+
+    /// @TODO send all this to template system
+    if(NULL==templateElement->export_fn) switch(templateElement->templateElementId)
+    {
+#ifdef HAVE_CISCO_URL
+      case CISCO_URL:
+        // do nothing. subtask will do.
+        break;
+
+      case CISCO_HTTP_HOST_L1:
+      case CISCO_HTTP_REFERER_L1:
+      if(real_field_len > 6){
+        char * referer = malloc(sizeof(char)*(real_field_len -6 + 1));
+        snprintf(referer,real_field_len-6,"%s",buffer+real_field_len_offset+6);
+        size_t domain_len;
+        const char *l1_domain = rb_l1_domain(referer,&domain_len, readOnlyGlobals.rb_databases.domains_name_as_list);
+        if(l1_domain)
+          value_ret = append_escaped(kafka_line_buffer,l1_domain,domain_len);
+        free(referer);
+      }
+      break;
+
+#if 0 /* SOCIAL_MEDIA */
+      case CISCO_HTTP_USER_AGENT_OS:
+      {
+        rb_keyval_list_t * iter;
+        for(iter=readOnlyGlobals.rb_databases.os_name_as_list;value_ret==0 && iter;iter=iter->next)
+        {
+          if(strcasestr(buf,iter->val))
+            value_ret = sprintbuf(kafka_line_buffer, "%s", iter->key);
+        }
+      }
+      break;
+      case CISCO_HTTP_SOCIAL_USER_FB:
+        if(unlikely(strstr(buf,"akamaihd.net") || strstr(buf,"ak.fbcdn.net"))){
+          size_t phid_len;
+          const char * phid = extract_fb_photo_id(url,strlen(url),
+                                                 buf,
+                                                 &phid_len);
+          if(phid)
+          {
+            value_ret = sprintbuf(kafka_line_buffer,"http://www.facebook.com/profile.php?id=");
+            printbuf_memappend_fast(kafka_line_buffer,phid,phid_len-1);
+            value_ret += phid_len - 1;
+            social_user_printed = 1;
+          }else
+            return(0);
+        }
+        break;
+      case CISCO_HTTP_SOCIAL_MEDIA_TT:
+        if(unlikely(strstr(buf,"twitter.com") &&
+                         strstr(url,"status")))
+        {
+          value_ret = sprintbuf(kafka_line_buffer,"%s%s",buf,url);
+        }
+        break;
+      case CISCO_HTTP_SOCIAL_USER_TT:
+        {
+          // Format: *twitter.com/USER/status/###
+          char * twitter_com_string = strstr(buf,"twitter.com");
+          char * status_string      = strstr(url,"status");
+          if(unlikely(twitter_com_string && status_string))
+          {
+            value_ret = sprintbuf(kafka_line_buffer,twitter_com_string + strlen("twitter.com"),status_string - twitter_com_string);
+          }
+
+          /* Format: *api.twitter.com*?screen_name=USER* */
+          if(real_field_len>6)
+          {
+            const struct counted_string host = {buffer+real_field_len_offset+6,real_field_len-6};
+            const struct counted_string counted_url  = {url,strlen(url)};
+
+            const struct counted_string tw_user = extract_tw_user(&counted_url,&host);
+            if(tw_user.string)
+            {
+              const size_t initial_size = kafka_line_buffer->bpos;
+              printbuf_memappend_fast(kafka_line_buffer,"https://twitter.com/",strlen("https://twitter.com/"));
+              printbuf_memappend_fast(kafka_line_buffer,tw_user.string,tw_user.len);
+              value_ret = kafka_line_buffer->bpos - initial_size;
+              //sprintf(kafka_line_buffer,tw_user.len+1,"%s",tw_user.string);
+            }
+          }
+        }
+        if(unlikely(strstr(buf,"twitter.com") && strstr(url,"status")))
+        {
+          value_ret = print_twitter_user(kafka_line_buffer,buf);
+        }
+        break;
+
+      case CISCO_HTTP_SOCIAL_USER_YT:
+        {
+          /* Format: gdata.youtube.com/feeds/api/users/USERNAME */
+          if(real_field_len > 6)
+          {
+            const struct counted_string host = {buffer+real_field_len_offset+6,real_field_len-6};
+            const struct counted_string counted_url  = {url,strlen(url)};
+
+            const struct counted_string yt_user = extract_yt_user(&host,&counted_url);
+            if(yt_user.len > 0 && yt_user.string)
+            {
+              const size_t initial_size = kafka_line_buffer->bpos;
+              printbuf_memappend_fast(kafka_line_buffer,"https://www.youtube.com/user/",strlen("https://www.youtube.com/user/"));
+              printbuf_memappend_fast(kafka_line_buffer,yt_user.string,yt_user.len);
+              value_ret = kafka_line_buffer->bpos - initial_size;
+            }
+          }
+        }
+        break;
+
+      case CISCO_HTTP_SOCIAL_USER_DROPBOX:
+      {
+        if(real_field_len>6)
+        {
+          const struct counted_string host = {buffer+real_field_len_offset+6,real_field_len-6};
+          const struct counted_string counted_url  = {url,strlen(url)};
+
+          const struct counted_string dropbox_user = extract_dropbox_user(&host,&counted_url);
+          if(dropbox_user.len > 0 && dropbox_user.string)
+          {
+            const size_t initial_size = kafka_line_buffer->bpos;
+            printbuf_memappend_fast(kafka_line_buffer,"Dropbox id: ",strlen("Dropbox id:"));
+            printbuf_memappend_fast(kafka_line_buffer,dropbox_user.string,dropbox_user.len);
+            value_ret = kafka_line_buffer->bpos - initial_size;
+          }
+        }
+      }
+      case CISCO_HTTP_SOCIAL_USER_YT_REFERER:
+        {
+          /* Format: gdata.youtube.com/feeds/api/users/USERNAME */
+          if(real_field_len > 6)
+          {
+            const struct counted_string referer = {buffer+real_field_len_offset+6,real_field_len-6};
+
+            const struct counted_string yt_user = extract_yt_user_referer(&referer);
+            if(yt_user.len > 0 && yt_user.string)
+            {
+              const size_t initial_size = kafka_line_buffer->bpos;
+              printbuf_memappend_fast(kafka_line_buffer,"https://www.youtube.com/user/",strlen("https://www.youtube.com/user/"));
+              printbuf_memappend_fast(kafka_line_buffer,yt_user.string,yt_user.len);
+              value_ret = kafka_line_buffer->bpos - initial_size;
+            }
+          }
+        }
+        break;
+      case CISCO_HTTP_SOCIAL_MEDIA_IG:
+        if(unlikely(NULL!=strstr(buf,"instagram")) && strstr(buf,"distilleryimage"))
+        {
+          value_ret = sprintbuf(kafka_line_buffer,"%s%s",buf,url);
+        }
+        break;
+      case CISCO_HTTP_SOCIAL_MEDIA_YT:
+        if(unlikely(NULL!=strstr(buf,"youtube") && (NULL!=strstr(url,"watch") || NULL!=strstr(url,"embed"))))
+        {
+          value_ret = sprintbuf(kafka_line_buffer,"%s%s",buf,url);
+        }
+        break;
 #endif
+#endif
+      default:
+        if (unlikely(!templateElement->export_fn &&
+                                        !templateElement->postTemplate &&
+                                        readOnlyGlobals.enable_debug)) {
+          traceEvent(TRACE_ERROR, "Unknown template id %s(%d).\n",
+            templateElement->jsonElementName,
+            templateElement->templateElementId);
+        }
+        break;
+    };
+
+    if(value_ret < 0) /* Error adding */
+    {
+      // traceEvent(TRACE_ERROR, "Cannot add value to kafka buffer.\n");
+    }
+    else if(value_ret > 0) /* Some added */
+    {
+      if(templateElement->quote)
+      {
+        value_ret+=2;
+        printbuf_memappend_fast(kafka_line_buffer,"\"",strlen("\""));
+      }
+    }
+    else /*if(value_ret == 0)*/
+    {
+      kafka_line_buffer->bpos = start_bpos;
+      kafka_line_buffer->buf[kafka_line_buffer->bpos] = '\0';
+    }
+
+    int i;
+    for(i=0; templateElement->postTemplate != NULL
+                            && templateElement->postTemplate[i] != NULL; ++i) {
+          printNetflowRecordWithTemplate(kafka_line_buffer,
+            templateElement->postTemplate[i], buffer, real_field_len,
+            real_field_len_offset, flowCache);
+    }
+
+    switch(templateElement->templateElementId) /* See if we can extract more information about the field */
+    {
+      // @TODO delete value_ret, not needed anymore.
+      #define print_if_template_valid(value_ret,template_elm)                                                            \
+      if(template_elm){                                                                                                  \
+        const int start_bpos = kafka_line_buffer->bpos;                                                                  \
+        const int ret = printNetflowRecordWithTemplate(kafka_line_buffer,template_elm,buffer,real_field_len,real_field_len_offset,flowCache); \
+        if(0==ret){                                                                                                      \
+          kafka_line_buffer->bpos = start_bpos;                                                                          \
+          kafka_line_buffer->buf[kafka_line_buffer->bpos] = '\0';                                                        \
+        }else{                                                                                                           \
+          value_ret+=ret;                                                                                                \
+        }                                                                                                                \
+      }
+
+      case CISCO_URL:
+        switch(buffer[real_field_len_offset+5]){
+          case 1:
+            //if(real_field_len>6 && buffer)
+            //  snprintf(url,min(sizeof(url),(unsigned)real_field_len-6+1),"%s",buffer+real_field_len_offset+6);
+            break;
+          case 2:
+            //print_if_template_valid(value_ret,readOnlyGlobals.rb_cached_templates.http_host_l1_template);
+            #if 1 /* MEDIA */
+            if(real_field_len>6)
+            {
+              //social_user_printed = 0; // <- It cannot be printed here.
+              //print_if_template_valid(value_ret,readOnlyGlobals.rb_cached_templates.http_social_media_tt_template);
+              //print_if_template_valid(value_ret,readOnlyGlobals.rb_cached_templates.http_social_user_fb_template);
+              //print_if_template_valid(value_ret,readOnlyGlobals.rb_cached_templates.http_social_user_tt_template);
+              //print_if_template_valid(value_ret,readOnlyGlobals.rb_cached_templates.http_social_user_yt_template);
+              //print_if_template_valid(value_ret,readOnlyGlobals.rb_cached_templates.http_social_user_dropbox);
+              //print_if_template_valid(value_ret,readOnlyGlobals.rb_cached_templates.http_social_media_ig_template);
+              //print_if_template_valid(value_ret,readOnlyGlobals.rb_cached_templates.http_social_media_yt_template);
+            }
+            //url[0]='\0';
+            break;
+          case 3:
+            // print_if_template_valid(value_ret,readOnlyGlobals.rb_cached_templates.http_user_agent_template);
+
+            break;
+          case 4:
+            // print_if_template_valid(value_ret,readOnlyGlobals.rb_cached_templates.http_referer_template);
+            //if(!social_user_printed)
+            //  print_if_template_valid(value_ret,readOnlyGlobals.rb_cached_templates.http_social_user_yt_referer);
+            #endif
+            break;
+          default:
+            // Do nothing
+            break;
+        }
+        break;
+        default:
+          // Do nothing
+          break;
+    };
+
+  } /* end value print if no error */
+  return value_ret;
+}
+
+/** Separate kafka flow message in many flows, one per minute, and distributing
+  the packets and octets proportionally.
+ * @param  kafka_line_buffer Current string with parameters that all flows share
+ * @param  first_timestamp   First timestamp of the flow
+ * @param  dSwitched         Delta time of flow.
+ * @param  dInterval         Maximum intervals allowed.
+ * @param  max_intervals     Maximum number of intervals to split
+ * @param  bytes             Total bytes
+ * @param  pkts              Total packets
+ * @return                   String list with all time separated flows
+ * @todo This should be in collect.c
+ */
+struct string_list *rb_separate_long_time_flow(
+  struct printbuf * kafka_line_buffer,
+  uint64_t first_timestamp, uint64_t dSwitched, const uint64_t dInterval,
+  const uint64_t max_intervals, uint64_t bytes, uint64_t pkts) {
+
+  struct string_list *kafka_buffers_list = NULL;
+
+  if(!kafka_line_buffer)
+    return kafka_buffers_list;
+
+  unsigned n_intervals_except_first = dSwitched/dInterval;
+  if(n_intervals_except_first>max_intervals){
+    if(unlikely(readOnlyGlobals.enable_debug))
+      traceEvent(TRACE_WARNING,
+        "Too many intervals to divide (%u, max: %"PRIu64"). nIntervals set to "
+        "%lu",
+        n_intervals_except_first, max_intervals, max_intervals-1);
+    n_intervals_except_first=max_intervals-1;
+  }
+
+  /* Avoiding flows with 0 pkts/bytes */
+  if(dSwitched > 0 && dSwitched % dInterval == 0)
+    dSwitched--;
+  const time_t first_interval_duration  = dSwitched%dInterval;
+  const double first_interval_duration_percent  = dSwitched == 0 ? 1 : (double)first_interval_duration/dSwitched;
+
+  const uint64_t first_interval_bytes  = bytes * first_interval_duration_percent;
+  const uint64_t first_interval_pkts   = pkts  * first_interval_duration_percent;
+
+  const uint64_t packets_per_interval    = n_intervals_except_first > 0 ? (pkts - first_interval_pkts)/n_intervals_except_first : 0;
+  const uint64_t bytes_per_interval      = n_intervals_except_first > 0 ? (bytes - first_interval_bytes)/n_intervals_except_first : 0;
+  const uint64_t packets_per_interval_sw = htonll(packets_per_interval);
+  const uint64_t bytes_per_interval_sw   = htonll(bytes_per_interval);
+
+  const uint64_t remainder_packets = pkts - first_interval_pkts - packets_per_interval*n_intervals_except_first;
+  const uint64_t remainder_bytes   = bytes - first_interval_bytes - bytes_per_interval*n_intervals_except_first;
+
+  const uint64_t first_interval_pkts_sw  = htonll(first_interval_pkts + remainder_packets);
+  const uint64_t first_interval_bytes_sw = htonll(first_interval_bytes + remainder_bytes);
+
+  uint64_t current_timestamp_s  = first_timestamp;
+
+  const char *last_sentmsg = NULL;
+  size_t last_sentmsg_useful_bpos = 0; /* bpos of last_sentmsg before timestamp -> useful*/
+
+  while(current_timestamp_s <= first_timestamp + dSwitched){
+    struct string_list *node = calloc(1,sizeof(struct string_list));
+    if(NULL==node){
+      traceEvent(TRACE_ERROR,"Memory error");
+      return kafka_buffers_list;
+    }
+
+    node->next = kafka_buffers_list?kafka_buffers_list:NULL;
+    kafka_buffers_list = node;
+    if(current_timestamp_s + dInterval > first_timestamp + dSwitched){
+      /* Last interval -> We use the buffer passed and remainder byes */
+      node->string = kafka_line_buffer;
+      printNetflowRecordWithTemplate(node->string, TEMPLATE_OF(PRINT_IN_BYTES),
+        (const char *)&first_interval_bytes_sw,
+        sizeof(first_interval_bytes_sw), 0, NULL);
+      printNetflowRecordWithTemplate(node->string, TEMPLATE_OF(PRINT_IN_PKTS),
+        (const char *)&first_interval_pkts_sw,
+        sizeof(first_interval_pkts_sw), 0, NULL);
+    }else{
+      /* Not last interval -> we need to clone the buffer */
+      node->string = printbuf_new();
+      if(NULL == node->string){
+        /* Can't allocate string: Return, at least, previous list */
+        kafka_buffers_list = node->next;
+        free(node);
+        return kafka_buffers_list;
+      }
+
+      if(NULL==last_sentmsg){
+        /* No cached message to copy from -> need to generate the first */
+        printbuf_memappend(node->string,kafka_line_buffer->buf,kafka_line_buffer->bpos);
+        printNetflowRecordWithTemplate(node->string, TEMPLATE_OF(PRINT_IN_BYTES),
+          (const char *)&bytes_per_interval_sw,
+          sizeof(bytes_per_interval_sw), 0, NULL);
+        printNetflowRecordWithTemplate(node->string, TEMPLATE_OF(PRINT_IN_PKTS),
+          (const char *)&packets_per_interval_sw,
+          sizeof(packets_per_interval_sw), 0, NULL);
+        last_sentmsg = node->string->buf;
+        last_sentmsg_useful_bpos = node->string->bpos;
+      }else{
+        /* Cached message: We could use it to waste less CPU */
+        printbuf_memappend(node->string,last_sentmsg,last_sentmsg_useful_bpos);
+      }
+    }
+
+    const uint64_t interval_timestamp_sw = htonll(current_timestamp_s);
+    printNetflowRecordWithTemplate(node->string, TEMPLATE_OF(PRINT_LAST_SWITCHED),
+      (const char *)&interval_timestamp_sw, sizeof(interval_timestamp_sw),
+      0, NULL);
+
+    current_timestamp_s += dInterval;
+
+    printbuf_memappend_fast(node->string,"}",strlen("}"));
+  }/* foreach interval in nIntervals */
+
+  return kafka_buffers_list;
+}
